# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Susu Collection System API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionPaymentMethod } from './collectionPaymentMethod';
import type { CollectionStatus } from './collectionStatus';

export interface Collection {
  id: number;
  customerId: number;
  customerName: string;
  amount: number;
  collectionDate: Date;
  paymentMethod: CollectionPaymentMethod;
  /** @nullable */
  notes?: string | null;
  status: CollectionStatus;
  /** @nullable */
  receiptId?: number | null;
  createdAt: Date;
}