# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Susu Collection System API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionInputPaymentMethod } from './collectionInputPaymentMethod';

export interface CollectionInput {
  customerId: number;
  /** @minimum 0.01 */
  amount: number;
  collectionDate: Date;
  paymentMethod: CollectionInputPaymentMethod;
  notes?: string;
}