# lib/api-zod/src/generated/types/collector.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Susu Collection System API
 * OpenAPI spec version: 0.1.0
 */

export interface Collector {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  zone: string;
  /** @nullable */
  avatarUrl?: string | null;
  totalCustomers?: number;
  createdAt?: Date;
}