# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Susu Collection System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalCollectedToday: number;
  assignedCustomers: number;
  pendingCollections: number;
  loanRepaymentsToday: number;
  totalCollectedMonth: number;
  /** Percentage of assigned customers collected today */
  collectionRate?: number;
}